- Reduce GCC 9 warnings.
git-svn-id: https://svn.code.sf.net/p/codeblocks/code/trunk@12668 2a5c6006-c6dd-42ca-98ab-0921f2732cef

+++ src/plugins/contrib/wxContribItems/wxled/ledpanel/src/wxledpanel.cpp
@@ -603,8 +603,8 @@ void wxLEDPanel::OnScrollTimer(wxTimerEvent& event)
     else
     {
         m_aniFrameNr++;
-        if(m_aniFrameNr >= m_ani.GetFrameCount())
-            m_aniFrameNr=0;
+        if (m_aniFrameNr >= (int)m_ani.GetFrameCount())
+            m_aniFrameNr = 0;
 
         m_content_mo.Init(m_ani.GetFrame(m_aniFrameNr));
         m_field.Clear();
